changed .node to .endpoint fixed

+++ source/RedisInside/TemporaryFile.cs
@@ -32,7 +32,8 @@ namespace PdfExtract
                 return;
             try
             {
-                _fileInfo.Delete();
+                if (disposing)
+                    _fileInfo.Delete();
             }
             catch (Exception ex)
             {
